requirements.txt
groq
python-dotenv
colorama
pytest
